日志表完成，关于请款预算单的录入日志完成

+++ src/main/java/com/cdkj/loan/ao/ISYSBizLogAO.java
@@ -18,6 +18,6 @@ public interface ISYSBizLogAO {
 
     public List<SYSBizLog> querySYSBizLogList(SYSBizLog condition);
 
-    public SYSBizLog getSYSBizLog(Long id);
+    public SYSBizLog getSYSBizLog(int id);
 
 }
